+++ apps/dash/urls.py
@@ -1,9 +1,11 @@
 from rest_framework.routers import DefaultRouter
 
-from .views import (
+from .views.views import (
     CarsView, SeatView, JourneyView, RoutingView, CoverCityView,
     PointOfSaleView, PointOfSaleWorkerView, CabinePlaneView, JourneyTarifView, JourneyClassView
 )
+
+from .views.routes import RouteProcessView
 
 router = DefaultRouter()
 router.register(r'technic/cars', CarsView, basename='cars')
@@ -13,6 +15,8 @@ router.register(r'technic/cabine_plane',
 router.register(r'transport/classe', JourneyClassView, basename='classes')
 router.register(r'transport/tarif', JourneyTarifView, basename='tarif')
 router.register(r'transport/journey', JourneyView, basename='journey')
+router.register(r'transport/routes',
+                RouteProcessView, basename='route')
 router.register(r'transport/routing', RoutingView, basename='routing')
 router.register(r'transport/cover_city', CoverCityView, basename='cover_city')
 router.register(r'transport/pos', PointOfSaleView, basename='point_of_sale')
